changed the structure of the constructors of BaseMapping and BaseAnalyticMapping, now to adjust it to SplineMappings in Psydac
